backend/pyproject.toml
[project]
name = "adpt-backend"
version = "0.1.0"
description = "FastAPI backend for adpt developer profiles."
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.0",
  "psycopg[binary]>=3.2.0",
  "pydantic-settings>=2.5.0",
  "python-jose[cryptography]>=3.3.0",
  "httpx>=0.27.0",
  "openai>=1.59.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.24.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

